requirements.txt
aiohttp==3.8.5
aiokafka==0.8.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
confluent-kafka==2.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.3
django-environ==0.11.2
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
environ==1.0
frozenlist==1.4.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.3
multidict==6.0.4
mysqlclient==2.2.0
openai==0.28.0
openapi-codec==1.3.2
packaging==23.1
pytz==2023.3.post1
requests==2.31.0
simplejson==3.19.1
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.5
yarl==1.9.2
